lcw
1.在求购里，如果求购描述字数太多在显示时，会超出屏幕下方的个人信息条。
2.Ios系统中修改密码后，不会自动跳入重新登录页面。而且如果此时发车，发布的车源内容页会出现联系人错误
3、我的收藏里没有发布日期和时间
4、发布车源里，添加照片，点击大图图片右上角无明显标识
5、我的车源、我的求购最底部信息操作，无法完成操作栏自动上弹

6、求购页个人信息跟着地区变化而变（接口需要添加用户地址）
7.商圈不显示新添加的好友。（接口问题）
8、我的车源、我的求购 排序（接口修改排序方式）
9、在个人中心，我的车源、我的求购对信息进行操作时，闪退。（暂定）
10、生产日期改为不实必须项（暂定）
11、颜色标签添加自定义（暂定）

+++ FBAuto/FBAuto/Public/TakePhotos/QBImagePickerController.m
@@ -249,7 +249,7 @@
     
     cell.imageView.image = [UIImage imageWithCGImage:assetsGroup.posterImage];
     cell.titleLabel.text = [NSString stringWithFormat:@"%@", [assetsGroup valueForProperty:ALAssetsGroupPropertyName]];
-    cell.countLabel.text = [NSString stringWithFormat:@"(%ld)", assetsGroup.numberOfAssets];
+    cell.countLabel.text = [NSString stringWithFormat:@"(%ld)", (long)assetsGroup.numberOfAssets];
     
     return cell;
 }
